requirements.txt
Flask
transformers
torch
torchvision
torchaudio
langchain
python-dotenv
openai